[#390] Fix issues with PR
* Merge `filter_technical_parameters_helper` and
  `parse_offer_parameter_value_helper` into `attributes_helper`

* Other small code style fixes

Fixes: #390
